+++ test/Map/StyledHtmlSpec.tsx
@@ -16,8 +16,7 @@ describe("StyledHtml", function () {
       baseUrl: "./"
     });
     viewState = new ViewState({
-      terria: terria,
-      catalogSearchProvider: undefined
+      terria: terria
     });
   });