+++ test/ReactViews/Map/Navigation/Compass/GyroscopeGuidanceSpec.tsx
@@ -13,8 +13,7 @@ describe("GyroscopeGuidance", function () {
       baseUrl: "./"
     });
     viewState = new ViewState({
-      terria: terria,
-      catalogSearchProvider: undefined
+      terria: terria
     });
   });