+++ test/ReactViews/Map/Panels/HelpPanel/HelpPanelSpec.tsx
@@ -14,8 +14,7 @@ describe("HelpPanel", function () {
       baseUrl: "./"
     });
     viewState = new ViewState({
-      terria: terria,
-      catalogSearchProvider: undefined
+      terria: terria
     });
   });