+++ test/ReactViews/SidePanel/BrandingSpec.tsx
@@ -11,8 +11,7 @@ describe("Branding", function () {
   beforeEach(function () {
     terria = new Terria();
     viewState = new ViewState({
-      terria,
-      catalogSearchProvider: undefined
+      terria
     });
   });