+++ test/ReactViews/Tour/TourPortalSpec.tsx
@@ -16,8 +16,7 @@ describe("TourPortal", function () {
       baseUrl: "./"
     });
     viewState = new ViewState({
-      terria: terria,
-      catalogSearchProvider: undefined
+      terria: terria
     });
   });